Follow the output format of exuberant-ctags
I'm trying to merge the output from ctags and that from
coffeetags. Here ctags is used for .c and .h files; and coffeetags for
.coffee files.

I found gaps between two tags files.

ctags uses "line" where "lineno" used in coffeetags.
ctags uses "kind" where "type" used in coffeetags.

In addition language field is missing in the output of coffeetags.
From the view of coffeetags, language field is non-sense. However, when
mixing the output, the field is helpful to know the origin of each tag line.

I hope the output of coffeetags follows that of ctags.

Here is an example of ctags output:

    % ctags -o - --fields=+lntzkK ../yacc.c
    YaccParser	../yacc.c	/^extern parserDefinition* YaccParser ()$/;"	kind:function	line:30	language:C
    installYaccRegex	../yacc.c	/^static void installYaccRegex (const langType language)$/;"	kind:function	line:24	language:C	...

Here is an example of coffeetags output without this patch:

    %  coffeetags --include-vars foo.coffee
    ...
    BTree	foo.coffee	/^BTree   = ->$/;"	f	lineno:191	object:window	type:function
    BTree	foo.coffee	/^BTree = ->$/;"	f	lineno:203	object:window	type:function
    add	foo.coffee	/^add = (x, y) -> x + y$/;"	f	lineno:11	object:window	type:function
    b	foo.coffee	/^b       = new BTree$/;"	o	lineno:192	object:window	type:object
    b	foo.coffee	/^b = new BTree$/;"	o	lineno:204	object:window	type:object

Here is an example of coffeetags output with this patch:

    % bash ./run.sh foo.coffee
    ...
    BTree	foo.coffee	/^BTree   = ->$/;"	f	line:191	object:window	kind:function	language:coffee
    BTree	foo.coffee	/^BTree = ->$/;"	f	line:203	object:window	kind:function	language:coffee
    add	foo.coffee	/^add = (x, y) -> x + y$/;"	f	line:11	object:window	kind:function	language:coffee
    b	foo.coffee	/^b       = new BTree$/;"	o	line:192	object:window	kind:object	language:coffee
    b	foo.coffee	/^b = new BTree$/;"	o	line:204	object:window	kind:object	language:coffee

If this patch breaks interchangeability with other tools, I will wirte a patch
adding --follow-ctags-output.

Signed-off-by: Masatake YAMATO <yamato@redhat.com>